Added Ui interactions

+++ app/src/main/java/com/example/aairastation/feature_menu/presentation/menu_list/MenuListViewModel.kt
@@ -34,7 +34,7 @@ class MenuListViewModel @Inject constructor(
             }
         }
 
-    fun viewItem(navigator: DestinationsNavigator, id: Int) {
+    fun viewItem(navigator: DestinationsNavigator, id: Long) {
 //        navigator.navigate(ForageItemDetailScreenDestination(id))
     }
